feat: add LinkManager component for managing links in settings
- Introduced LinkManager component to handle adding, editing, and removing links.
- Integrated LinkManager into SettingInput for specific settings (footer_links, header_links).
- Updated package.json to include react-beautiful-dnd and its types for drag-and-drop functionality.
- Refactored SettingsIndex to utilize SettingInput for rendering different setting types.
- Improved JSON validation and formatting for settings.
